+++ gate-web/src/main/groovy/com/netflix/spinnaker/gate/services/internal/OpsmxOesService.groovy
@@ -195,4 +195,7 @@ interface OpsmxOesService {
   Object evaluateStaticPolicy(@Path('version') String version,
                               @Body Object data)
 
+  @POST("/oes/accEnvMapping/import")
+  Object importAccountEnvironmentMappings()
+
 }
